splashscreen added
